Change C++ ops based on reviewer's comment
